pyproject.toml
[project]
name = "doc-env"
version = "0.1.0"
description = "Hospital Scheduler RL Environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic",
    "gymnasium>=1.1.1",
    "stable-baselines3>=2.7.1",
    "google-generativeai>=0.8.6",
    "google-genai>=1.47.0",
    "openai>=2.30.0",
    "httpx>=0.28.1",
]

[project.scripts]
doc-env-server = "server.app:main"
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [".", "server"]
